moving function to utils.py
