+++ custom_code/cadences/snex_resume_cadence_after_failure.py
@@ -8,31 +8,10 @@ from tom_observations.models import ObservationRecord
 
 from tom_observations.cadences.resume_cadence_after_failure import ResumeCadenceAfterFailureStrategy
 
-from custom_code.utils import format_form_errors, sync_group_permissions_to_target
+from custom_code.utils import apply_proposal_rollover, format_form_errors, sync_group_permissions_to_target
 
 
 logger = logging.getLogger(__name__)
-
-
-def apply_proposal_rollover(observation_payload, start_keyword='start'):
-    start = None
-    for rollover in getattr(settings, 'PROPOSAL_ROLLOVERS', []):
-        if observation_payload.get('proposal') != rollover['old_id']:
-            continue
-        if start is None:
-            start_value = observation_payload.get(start_keyword)
-            if not start_value:
-                return observation_payload
-            start = parse(start_value) if isinstance(start_value, str) else start_value
-            if timezone.is_naive(start):
-                start = timezone.make_aware(start)
-        semester_start = parse(rollover['semester_start'])
-        if timezone.is_naive(semester_start):
-            semester_start = timezone.make_aware(semester_start)
-        if start >= semester_start:
-            logger.info(f"Rolling over proposal {rollover['old_id']} to {rollover['new_id']} for window starting {observation_payload.get(start_keyword)}")
-            observation_payload['proposal'] = rollover['new_id']
-    return observation_payload
 
 
 class SnexCadencePermissionMixin:
